security fix, user moderation improvement
Added a fix to prevent IP spoofing when using cloudflare (user needs to specify if cloudflare is enabeld in the config file)

Also, when viewing images uploaded by an IP address, images might not have thumbs if the user wasn't logged in. thumbnails will now be created when moderating a user. this can be enabled/disabled in the config file

Unfortunately this meant removing support for additional file extensions for now as it conflicted with with the thumbnail creation (something I missed before), but this may be re-added in the future